refactor: 모듈 네이밍 변경
