+++ Projects/Feature/Profile/Project.swift
@@ -2,59 +2,44 @@ import ProjectDescription
 import ProjectDescriptionHelpers
 import DependencyPlugin
 
-let project = Project.makeModule(
-    name: ModulePath.Feature.name+ModulePath.Feature.Profile.rawValue,
-    targets: [    
-        .feature(
-            interface: .Profile,
-            factory: .init(
-                dependencies: [
-                    .domain
-                ]
-            )
-        ),
-        .feature(
-            implements: .Profile,
-            factory: .init(
-                dependencies: [
-                    .feature(interface: .Profile)
-                ]
-            )
-        ),
-        .feature(
-            testing: .Profile,
-            factory: .init(
-                dependencies: [
-                    .feature(interface: .Profile)
-                ]
-            )
-        ),
-        .feature(
-            tests: .Profile,
-            factory: .init(
-                dependencies: [
-                    .feature(testing: .Profile)
-                ]
-            )
-        ),
-        .feature(
-            example: .Profile,
-            factory: .init(
-                infoPlist: .extendingDefault(
-                    with: [
-                        "CFBundleShortVersionString": "1.0",
-                        "CFBundleVersion": "1",
-                        "UILaunchStoryboardName": "LaunchScreen",
-                        "UIApplicationSceneManifest": [
-                            "UIApplicationSupportsMultipleScenes": false,
-                            "UISceneConfigurations": []
-                        ]
-                    ]),
-                dependencies: [
-                    .feature(interface: .Profile),
-                    .feature(implements: .Profile)
-                ]
-            )
+let targets: [Target] = [
+    .feature(
+        implements: .Profile,
+        factory: .init(
+            dependencies: [
+                .feature(interface: .Home),
+                .feature(interface: .Profile),
+                .feature(interface: .Onboarding),
+            ]
         )
-    ]
+    ),
+    .feature(
+        tests: .Profile,
+        factory: .init(
+            dependencies: [
+                .feature(testing: .Profile)
+            ]
+        )
+    ),
+    .feature(
+        testing: .Profile,
+        factory: .init(
+            dependencies: [
+                .feature(interface: .Profile)
+            ]
+        )
+    ),
+    .feature(
+        interface: .Profile,
+        factory: .init(
+            dependencies: [
+                .domain
+            ]
+        )
+    )
+]
+
+let project: Project = .makeModule(
+    name: "FeatureProfile",
+    targets: targets
 )
