+++ Projects/Feature/Project.swift
@@ -17,7 +17,6 @@ let targets: [Target] = [
                 .feature(implements: .Onboarding),
                 .feature(implements: .Home),
                 .feature(implements: .Profile),
-                .feature(implements: .MyPage),
                 .feature(implements: .Workout)
                 // make module을 통해 생성한 모듈을 각 레이어의 최상위부분에서 디펜던시를 주입시켜야함
             ]
